+++ water_balance/views/waterbalance_widget.py
@@ -660,6 +660,8 @@ class WaterBalanceWidget(QDockWidget):
             self.hover_enter_map_visualization)
         self.wb_item_table.hoverExitAllRows.connect(
             self.hover_exit_map_visualization)
+        self.activate_all_button.clicked.connect(self.activate_layers)
+        self.deactivate_all_button.clicked.connect(self.deactivate_layers)
 
         # TODO: is this a good default?
         # initially turn on tool
@@ -997,8 +999,13 @@ class WaterBalanceWidget(QDockWidget):
             self.select_polygon_button.setText(_translate(
                 "DockWidget", "Draw new polygon", None))
 
-    def redraw_wb(self):
-        pass
+    def activate_layers(self):
+        for item in self.model.rows:
+            item.active.value = True
+
+    def deactivate_layers(self):
+        for item in self.model.rows:
+            item.active.value = False
 
     def get_modelpart_graph_layers(self, graph_layers):
         modelpart_graph_series = [
@@ -1013,9 +1020,12 @@ class WaterBalanceWidget(QDockWidget):
 
         self.model.removeRows(0, len(self.model.rows))
         self.model.ts = ts
+
+        # self.layers_in_table = self.get_modelpart_graph_layers(
+        #     graph_series['items'])
+
         self.model.insertRows(
             self.get_modelpart_graph_layers(graph_series['items']))
-
         if self.agg_combo_box.currentText() == 'm3/s':
             self.plot_widget.setLabel("left", "flow", "m3/s")
         elif self.agg_combo_box.currentText() == 'm3 cumulative':
@@ -1250,6 +1260,8 @@ class WaterBalanceWidget(QDockWidget):
             self.hover_enter_map_visualization)
         self.wb_item_table.hoverExitAllRows.disconnect(
             self.hover_exit_map_visualization)
+        self.activate_all_button.clicked.disconnect(self.activate_layers)
+        self.deactivate_all_button.clicked.disconnect(self.deactivate_layers)
 
         self.closingWidget.emit()
         event.accept()
@@ -1289,11 +1301,22 @@ class WaterBalanceWidget(QDockWidget):
         self.agg_combo_box = QComboBox(self)
         self.button_bar_hlayout.addWidget(self.agg_combo_box)
 
+        # now first add a QSpacerItem so that the QPushButton (added sub-
+        # sequently) are aligned on the right-side of the button_bar_hlayout
         spacer_item = QSpacerItem(40,
                                   20,
                                   QSizePolicy.Expanding,
                                   QSizePolicy.Minimum)
         self.button_bar_hlayout.addItem(spacer_item)
+
+        self.activate_all_button = QPushButton(self)
+        self.button_bar_hlayout.addWidget(
+            self.activate_all_button, alignment=Qt.AlignRight)
+
+        self.deactivate_all_button = QPushButton(self)
+        self.button_bar_hlayout.addWidget(
+            self.deactivate_all_button, alignment=Qt.AlignRight)
+
         self.main_vlayout.addLayout(self.button_bar_hlayout)
 
         # add tabWidget for graphWidgets
@@ -1339,3 +1362,7 @@ class WaterBalanceWidget(QDockWidget):
             "DockWidget", "Show total balance", None))
         self.reset_waterbalans_button.setText(_translate(
             "DockWidget", "Hide on map", None))
+        self.activate_all_button.setText(_translate(
+            "DockWidget", "activate all", None))
+        self.deactivate_all_button.setText(_translate(
+            "DockWidget", "deactivate all", None))